change : internal spec (NOverlaySender, NOverlayHandler)
