+++ lib/flutter_naver_map.dart
@@ -17,14 +17,26 @@ import 'package:flutter_naver_map/src/util/image_util.dart';
 import 'package:flutter_naver_map/src/util/math.dart';
 import 'package:flutter_naver_map/src/util/widget_to_image.dart';
 
-part 'src/controller/map/controller.dart';
-part 'src/controller/map/handler.dart';
 /*
   --- controller ---
 */
+part 'src/controller/map/controller.dart';
+
+part 'src/controller/map/handler.dart';
+
 part 'src/controller/map/sender.dart';
+
+/*
+  --- OverlayController ---
+*/
 part 'src/controller/overlay/overlay_controller.dart';
+
 part 'src/controller/overlay/overlay_controller_impl.dart';
+
+part 'src/type/map/overlay/overlay/overlay_handler.dart';
+
+part 'src/type/map/overlay/overlay/overlay_sender.dart';
+
 /*
   --- exceptions ---
 */
@@ -34,47 +46,81 @@ part 'src/exceptions/exceptions.dart';
 */
 
 part 'src/initializer/flutter_naver_map_sdk_initializer.dart';
+
 part 'src/initializer/flutter_naver_map_sdk_initializer_impl.dart';
 /*
   --- type ---
 */
 
 part 'src/type/default/locale.dart';
+
 part 'src/type/default/padding.dart';
+
 part 'src/type/default/point.dart';
+
 part 'src/type/default/size.dart';
+
 part 'src/type/map/camera/camera_position.dart';
+
 part 'src/type/map/camera/camera_update.dart';
+
 part 'src/type/map/enums.dart';
+
 part 'src/type/map/geo/latlng.dart';
+
 part 'src/type/map/geo/latlng_bounds.dart';
+
 part 'src/type/map/indoor/indoor_level.dart';
+
 part 'src/type/map/indoor/indoor_region.dart';
+
 part 'src/type/map/indoor/indoor_zone.dart';
+
 part 'src/type/map/indoor/selected_indoor.dart';
+
 part 'src/type/map/info/overlay_info.dart';
+
 part 'src/type/map/info/overlay_query.dart';
+
 part 'src/type/map/info/pickable_info.dart';
+
 part 'src/type/map/info/symbol_info.dart';
+
 part 'src/type/map/map_options.dart';
+
 part 'src/type/map/overlay/overlay/addable/addable.dart';
+
 part 'src/type/map/overlay/overlay/addable/circle_overlay.dart';
+
 part 'src/type/map/overlay/overlay/addable/ground_overlay.dart';
+
 part 'src/type/map/overlay/overlay/addable/info_window.dart';
+
 part 'src/type/map/overlay/overlay/addable/marker.dart';
+
 part 'src/type/map/overlay/overlay/addable/path_overlay/arrow_head_path_overlay.dart';
+
 part 'src/type/map/overlay/overlay/addable/path_overlay/multipart_path.dart';
+
 part 'src/type/map/overlay/overlay/addable/path_overlay/multipart_path_overlay.dart';
+
 part 'src/type/map/overlay/overlay/addable/path_overlay/path_overlay.dart';
+
 part 'src/type/map/overlay/overlay/addable/polygon_overlay.dart';
+
 part 'src/type/map/overlay/overlay/addable/polyline_overlay.dart';
+
 part 'src/type/map/overlay/overlay/location_overlay.dart';
+
 part 'src/type/map/overlay/overlay/overlay.dart';
+
 part 'src/type/map/overlay/overlay_caption.dart';
+
 part 'src/type/map/overlay/overlay_image.dart';
 /*
   --- widget ---
 */
 
 part 'src/widget/map_widget.dart';
+
 part 'src/widget/platform_view.dart';
